sql/schemachanger: avoid retries on invalid expressions during backfills
Currently, the schema changer uses a white list to determine which
errors to be retired, which in some cases use a text comparison.
Unfortunately, when we parse expressions their text can appear in the
error and cause errors to be erroneously retried for backfills. To
address this, this patch explicitly wraps errors from eval.Expr with
SchemaChangerUserError to block retries.

Fixes: #141352

Release note (bug fix): Invalid default expressions could cause
backfilling schema changes to retry forever